anapback/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.1
django-bootstrap-daterangepicker==1.1.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
filelock==3.13.1
idna==3.6
platformdirs==4.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.1.0
virtualenv==20.25.1
